Introduce zend_debug_backtrace_as_string()
Refactor code from Exception#getTraceAsString()
into Zend/zend_builtin_functions.c as PHP_API and
reuse from zend_exceptions.c

This is the first step towards a new PHP function
debug_backtrace_string(); which behaves the same as
Exception#getTraceAsString().

This is helpful for logging and for extensions that
work with errors.
